+++ R/rte_Utils.R
@@ -59,6 +59,7 @@ rte.check.my.os <-function(){
 #'
 #' @param f.in The location and name of latex file
 #' @param pdf.dir.out The name of the folder for the output pdf
+#' @param do.clean.up Clean (delete) auxiliary latex files? (TRUE or NOT)
 #' @return A flag, TRUE if the latex compilation was a sucess and FALSE if not
 #' @examples
 #' f.in <- system.file("extdata", "MyRandomTest.tex", package = "RndTexExams")
@@ -69,7 +70,8 @@ rte.check.my.os <-function(){
 #'
 #' @export
 rte.compile.latex <- function(f.in,
-                              pdf.dir.out = 'PdfOut'){
+                              pdf.dir.out = 'PdfOut',
+                              do.clean.up = T){
 
   #require(tools)
 
@@ -79,28 +81,52 @@ rte.compile.latex <- function(f.in,
     stop('Cant find pdflatex.exe. Make sure you have a flavor of LaTeX installed in your OS.')
   }
 
-  my.latex.flavor <- rte.check.latex.flavor()
+  # compile latex using system command (OLD code, version 1.1.1)
+  # KEEP IT FOR FUTURE REFERENCE
 
-  if (my.latex.flavor =='miktex'){
-    # for miktex, use -quiet
-    my.c<- sprintf('pdflatex -quiet -interaction=batchmode -output-directory="%s" "%s"',pdf.dir.out, f.in)
-  }
+  # my.latex.flavor <- rte.check.latex.flavor()
+  #
+  # if (my.latex.flavor =='miktex'){
+  #   # for miktex, use -quiet
+  #   my.c<- sprintf('pdflatex -quiet -interaction=batchmode -output-directory="%s" "%s"',pdf.dir.out, f.in)
+  # }
+  #
+  # if (any(my.latex.flavor ==c('texlive','mactex','unknown'))){
+  #   # for texlive, mactex or unknown, use only batchmode
+  #   my.c <- sprintf('pdflatex -interaction=batchmode -output-directory="%s" "%s"', pdf.dir.out, f.in)
+  # }
+  #
+  # if (my.os == 'Windows'){
+  #   # use show.output.on.console for Windows
+  #   system(command = my.c, show.output.on.console = F)
+  # } else {
+  #   system(command = my.c)
+  # }
 
-  if (any(my.latex.flavor ==c('texlive','mactex','unknown'))){
-    # for texlive, mactex or unknown, use only batchmode
-    my.c <- sprintf('pdflatex -interaction=batchmode -output-directory="%s" "%s"', pdf.dir.out, f.in)
-  }
+  # compile using tools::texi2pdf (new code in version 1.2)
 
-  if (my.os == 'Windows'){
-    # use show.output.on.console for Windows
-    system(command = my.c, show.output.on.console = F)
-  } else {
-    system(command = my.c)
-  }
+  tools::texi2pdf(file = f.in, clean = do.clean.up, quiet = T)
 
   # remove path and extension for checking pdf
 
   file.name.noext <- basename(tools::file_path_sans_ext(f.in))
+
+  # check if folder exists
+
+  if (!dir.exists(pdf.dir.out)){
+    warning(paste('Folder', pdf.dir.out, 'does not exists. Creating it...'))
+    dir.create(pdf.dir.out)
+  }
+
+  # copy pdf files to folder
+
+  f.out.pdf <- paste0(file.name.noext,'.pdf')
+
+  file.copy(from = f.out.pdf, to = pdf.dir.out )
+
+  # delete pdf files
+
+  file.remove(f.out.pdf)
 
   # Check if pdf file exists
   if (!file.exists(paste0(pdf.dir.out, '/', file.name.noext, '.pdf'))){